+++ src/evaluator/identifier.go
@@ -2,7 +2,7 @@ package evaluator
 
 import (
 	"github.com/duwa-lang/duwa/src/ast"
-	"github.com/duwa-lang/duwa/src/modules/all"
+	"github.com/duwa-lang/duwa/src/modules"
 	"github.com/duwa-lang/duwa/src/object"
 )
 
@@ -10,7 +10,7 @@ func evalIdentifier(
 	node *ast.Identifier,
 	env *object.Environment,
 ) object.Object {
-	if libraryFunction, ok := all.Functions[node.Value]; ok {
+	if libraryFunction, ok := modules.Functions[node.Value]; ok {
 		return libraryFunction
 	}
